complete services

+++ project/src/main/java/CourseRegistration/project/Repository/CourseOfferingRepository.java
@@ -3,5 +3,5 @@ package CourseRegistration.project.Repository;
 import CourseRegistration.project.domain.CourseOffering;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CourseOfferingRepository extends JpaRepository<CourseOffering,Integer> {
+public interface CourseOfferingRepository extends JpaRepository<CourseOffering,Long> {
 }
