+++ project/src/main/java/CourseRegistration/project/Repository/RegistrationEventRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RegistrationEventRepository extends JpaRepository<RegistrationEvent,Integer> {
+public interface RegistrationEventRepository extends JpaRepository<RegistrationEvent,Long> {
 }
